requirements.txt
hikari[speedups]==2.0.0.dev102
hikari-lightbulb==1.3.1
uvloop
requests
geopy